requirements.txt
accelerate==0.30.1
datasets==2.21.0
diffusers==0.30.3
face_alignment==1.4.1
matplotlib==3.10.3
numpy==2.3.2
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
Pillow==11.3.0
scikit_learn==1.7.1
scipy==1.16.1
timm==1.0.19
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
transformers==4.41.0
~orch==2.4.0
